refactor: system debloat, remove ccvars
